CONCF-799 | combine Mercury base page objects
